requirements.txt
feedparser==6.0.11
requests==2.32.3
python-dateutil==2.9.0.post0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
jinja2==3.1.4
flask==3.0.3
python-dotenv==1.0.1
reportlab==4.2.5
tqdm==4.67.1
pyyaml==6.0.2
